Fix pets crashes (#226)
* fix some stuff maybe

* Boulberry Bushing

* spam jani logs

* maybe fix sob

* more spam

* does dat fix anytihnkgMjl

* more spam (real)

* ae
